ctf_playground/pyproject.toml
[project]
name = "ctf-playground"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pwntools>=4.14.1",
    "rpyc>=6.0.2",
]
